added eligibility check call
