+++ apps/main-landing/src/app/check-eligibility/types.ts
@@ -1,3 +1,26 @@
 export type FormPayload = {
-    address: string;
-  };
+  address: string;
+};
+
+export type EligibilityCheckResponse = {
+  "paid": number,
+  "free": number,
+  "extension": number,
+  "registration": string,
+  "invites": number,
+  "farcaster": number,
+  "hacker": number,
+  "ardent": number,
+  "aavegotchi": number,
+  "gitcoin": number,
+  "across": number,
+  "polymarket": number,
+  "snapshot": number,
+  "tally": number,
+  "jumper": number,
+  "time_multiplier": number,
+  "invite_multiplier": number,
+  "allocation_usage": number,
+  "allocation_partner": number,
+  "allocation": number
+}
